Examine if SampleReaderComp has created socket or not at daq_stop().
There are cases when we put a code that send slow control packet
at daq_start() AND throw fatal_error_report() if the slow controk fails.
In this case, the SampleReaderComp does not have a socket, so if we
transit to daq_stop() and delete m_sock, deleting will fail.  This commit
fixes this case.

This commit fixes #107 .